orchestration/requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-amazon==8.7.1